pyproject.toml
[tool.poetry]
name = "ai-agent-with-cursor"
version = "0.1.0"
description = "AI Agent для автоматизации анализа отклонений в Google Таблицах"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ai_agent", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"
google-api-python-client = "^2.108.0"
openai = "^1.3.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
requests = "^2.31.0"
pandas = "^2.1.0"
numpy = "^1.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
pre-commit = "^3.5.0"

[tool.poetry.scripts]
setup-google = "ai_agent.setup.google_setup:main"
proposals-from-drive = "ai_agent.jobs.proposals_from_drive:main"
test-connections = "ai_agent.jobs.proposals_from_drive:test_connection"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["ai_agent"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
